requirements.txt
discord.py-message-components==1.7.5.4
google_api_python_client==2.61.0
requests==2.28.1
